refactor: перемещает keks-db в feature-слой
