+++ js/feature/keks-db/index.js
@@ -1,6 +1,6 @@
-import { createUserDocument } from '../entity/user/document.js';
-import { createCommentDocument } from '../entity/comment/document.js';
-import { createPhotoDocument } from '../entity/photo/document.js';
+import { createUserDocument } from '../../entity/user/document.js';
+import { createCommentDocument } from '../../entity/comment/document.js';
+import { createPhotoDocument } from '../../entity/photo/document.js';
 import {
   APP_PHOTO_COUNT,
   APP_USERS_COUNT,
